1. Изменен алгоритм отображения строки на виртуальном дисплее 2. Переработан интерфейс дисплея 3. Добавлен модуль поддержки языков 4. Добавлена поддержка русско-английского набора символов 5. Переименованы файлы

+++ _Qt/project/tests/test_text_buffer.cpp
@@ -12,7 +12,7 @@ TestTextBuffer::TestTextBuffer(const QString & fileName, size_t maxSize)
     {
         a = f.readAll();
         a.remove(0, 2);
-        if(a.size() <= maxSize)
+        if(a.size() <= (int)maxSize)
         {
             size_t appSize = maxSize - a.size();
             a.append(appSize, 0);
